chore: reorganize test files into logical folders
- Created structured test directories mirroring source layout
- Organized into commands, services, bots, webhooks, utils, and integration
- Keeps Logger.test.ts in services folder
- Improves test discoverability and maintenance
