+++ src/__tests__/starbunk/reply-bots/blueBot.test.ts
@@ -1,8 +1,8 @@
+import { createMockMessage } from '@/__tests__/mocks/discordMocks';
+import { createMockWebhookService } from '@/__tests__/mocks/serviceMocks';
 import userID from '@/discord/userID';
 import { Logger } from '@/services/Logger';
 import BlueBot from '@/starbunk/bots/reply-bots/blueBot';
-import { createMockMessage } from '@/tests/mocks/discordMocks';
-import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import { Collection, GuildMember, Message, TextChannel, User } from 'discord.js';
 
 // Mock the OpenAIClient module
