+++ src/__tests__/starbunk/reply-bots/holdBot.test.ts
@@ -1,6 +1,6 @@
+import { createMockGuildMember, createMockMessage } from '@/__tests__/mocks/discordMocks';
+import { createMockWebhookService } from '@/__tests__/mocks/serviceMocks';
 import HoldBot from '@/starbunk/bots/reply-bots/holdBot';
-import { createMockGuildMember, createMockMessage } from '@/tests/mocks/discordMocks';
-import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import { Message, User } from 'discord.js';
 
 describe('HoldBot', () => {
