+++ src/__tests__/starbunk/reply-bots/pickleBot.test.ts
@@ -1,7 +1,7 @@
+import { createMockGuildMember, createMockMessage } from '@/__tests__/mocks/discordMocks';
+import { createMockWebhookService } from '@/__tests__/mocks/serviceMocks';
 import UserID from '@/discord/userID';
 import PickleBot from '@/starbunk/bots/reply-bots//pickleBot';
-import { createMockGuildMember, createMockMessage } from '@/tests/mocks/discordMocks';
-import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import Random from '@/utils/random';
 import { Message, User } from 'discord.js';
