+++ src/__tests__/starbunk/reply-bots/sheeshBot.test.ts
@@ -1,6 +1,6 @@
+import { createMockGuildMember, createMockMessage } from '@/__tests__/mocks/discordMocks';
+import { createMockWebhookService } from '@/__tests__/mocks/serviceMocks';
 import SheeshBot from '@/starbunk/bots/reply-bots/sheeshBot';
-import { createMockGuildMember, createMockMessage } from '@/tests/mocks/discordMocks';
-import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import { Message, User } from 'discord.js';
 
 describe('SheeshBot', () => {
